🎨 style(SideBar.tsx): replace div with Box component 🚀 chore(package.json): downgrade @mui/material to version 5.6.0 📝 docs(index.html): change title to E-Shop The Box component is used instead of a div to improve consistency with the Material UI library. The downgrade of the @mui/material package to version 5.6.0 is done to fix compatibility issues with other packages. The title of the HTML page is changed to E-Shop to reflect the name of the application.
🎨 style(main.tsx): import ThemeProvider and createTheme from @mui/material
The import of ThemeProvider and createTheme from @mui/material has been added to the import statements. This is to enable the use of Material UI's theming capabilities in the application.

🎨 style(Home.tsx): add MUI theme to Home page
The MUI theme is added to the Home page, which improves the consistency of the application's design. The theme is defined in the file and imported into the Home component. The NavBar component is wrapped in a ThemeProvider component, which applies the theme to all child components. The Box component is used to center the Outlet component and apply styles to the Home page.